pyproject.toml
[tool.poetry]
name = "burton"
version = "0.1.0"
description = ""
authors = ["Gary Ocean"]
license = "MIT"
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.10"
flask = "*"
flask-cors = "*"
python-dotenv = "*"
langchain = "^0.2.16"
langchain-openai = "^0.1.23"
langchain-community = "^0.2.16"
langchain-anthropic = "^0.1.23"
langchain-groq = "*"
langchain-huggingface = "*"
langchain-google-genai = "*"
openai = "*"
anthropic = "*"
groq = "*"
huggingface-hub = "*"
google-generativeai = "*"
beautifulsoup4 = "*"
pre-commit = "^3.8.0"
duckduckgo-search = "^6.2.11"
chromadb = "^0.5.5"
paramiko = "^3.4.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
